Update Logger to handle None for stdout stream
